Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -71,6 +71,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {
+            abort(403,"Access denied");
+        }
         return view("questions.edit",compact('question'));
     }
 
@@ -88,6 +91,9 @@ class QuestionsController extends Controller
             'body' => $request->body
             
         ]);*/
+        if(\Gate::denies('update-question', $question)) {
+            abort(403,"Access denied");
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success', "YOur question has been update");
     }
@@ -100,8 +106,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
+        if(\Gate::denies('delete-question', $question)) {
+            abort(403,"Access denied");
+        }
 
+        $question->delete();
         return redirect('questions')->with('success','Your question has been delete');
     }
 }
